@@ -17,33 +17,33 @@
     "url": "https://github.com/ioBroker/ioBroker.network"
   },
   "dependencies": {
-    "@iobroker/adapter-core": "^2.5.1",
+    "@iobroker/adapter-core": "^2.6.8",
     "netmask": "^2.0.2",
     "node-wifi": "^2.0.16",
-    "systeminformation": "^5.9.9"
+    "systeminformation": "^5.18.5"
   },
   "devDependencies": {
-    "@iobroker/testing": "^2.5.2",
-    "@types/chai": "^4.2.22",
-    "@types/chai-as-promised": "^7.1.4",
-    "@types/gulp": "^4.0.9",
-    "@types/mocha": "^9.0.0",
-    "@types/node": "^16.11.7",
+    "@iobroker/testing": "^4.1.0",
+    "@types/chai": "^4.3.5",
+    "@types/chai-as-promised": "^7.1.5",
+    "@types/gulp": "^4.0.11",
+    "@types/mocha": "^10.0.1",
+    "@types/node": "^20.3.1",
     "@types/proxyquire": "^1.3.28",
-    "@types/sinon": "^10.0.6",
-    "@types/sinon-chai": "^3.2.5",
-    "axios": "^0.24.0",
-    "chai": "^4.3.4",
+    "@types/sinon": "^10.0.15",
+    "@types/sinon-chai": "^3.2.9",
+    "axios": "^1.4.0",
+    "chai": "^4.3.7",
     "chai-as-promised": "^7.1.1",
-    "eslint": "^7.11.0",
+    "eslint": "^8.43.0",
     "gulp": "^4.0.2",
-    "mocha": "^9.1.3",
+    "mocha": "^10.2.0",
     "proxyquire": "^2.1.3",
-    "sinon": "^12.0.1",
+    "sinon": "^15.2.0",
     "sinon-chai": "^3.7.0",
-    "@alcalzone/release-script": "^3.4.1",
-    "@alcalzone/release-script-plugin-iobroker": "^3.4.1",
-    "@alcalzone/release-script-plugin-license": "^3.4.1"
+    "@alcalzone/release-script": "^3.5.9",
+    "@alcalzone/release-script-plugin-iobroker": "^3.5.9",
+    "@alcalzone/release-script-plugin-license": "^3.5.9"
   },
   "os": [
     "linux"
@@ -68,7 +68,9 @@
     "lint": "eslint",
     "build": "gulp",
     "release": "release-script patch --yes",
-    "release-minor": "release-script minor --yes"
+    "release-patch": "release-script patch --yes",
+    "release-minor": "release-script minor --yes",
+    "release-major": "release-script major --yes",    "update-packages": "ncu --upgrade && cd src && ncu --upgrade && cd .."
   },
   "bugs": {
     "url": "https://github.com/ioBroker/ioBroker.network/issues"
