@@ -3,20 +3,27 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@iobroker/adapter-react": "^2.0.14",
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/icons": "^4.11.2",
-    "@sentry/browser": "^6.14.1",
-    "@sentry/integrations": "^6.14.1",
-    "clsx": "^1.1.1",
+    "@iobroker/adapter-react-v5": "^4.1.0",
+    "@mui/material": "^5.13.5",
+    "@mui/icons-material": "^5.11.16",
+    "@mui/styles": "^5.13.2",
+    "@babel/eslint-parser": "^7.22.5",
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
+    "@sentry/browser": "^7.55.2",
+    "@sentry/integrations": "^7.55.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-icons": "^4.9.0",
+    "react-scripts": "^5.0.1",
+    "notistack": "^3.0.1",
     "eslint-plugin-eqeqeq-fix": "^1.0.3",
-    "eslint-plugin-only-warn": "^1.0.3",
-    "moment": "^2.29.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-icons": "^4.3.1",
-    "react-scripts": "^4.0.3",
-    "notistack": "^1.0.10"
+    "eslint-plugin-only-warn": "^1.1.0",
+    "eslint": "^8.43.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-jsx-a11y": "^6.7.1",
+    "eslint-plugin-react": "^7.32.2",
+    "eslint-plugin-react-hooks": "^4.6.0"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -36,11 +43,6 @@
     "not op_mini all"
   ],
   "devDependencies": {
-    "eslint": "^7.32.0",
-    "eslint-config-airbnb": "^18.2.1",
-    "eslint-plugin-import": "^2.24.2",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-react": "^7.25.1",
-    "eslint-plugin-react-hooks": "^4.2.0"
+
   }
 }
